requirements.txt
absl-py==1.0.0

dj-database-url==0.5.0
Django==4.0.3
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
django-model-utils==4.2.0

dnspython==2.2.1

djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0

GitPython==3.1.27
google-auth-oauthlib==0.4.6

gunicorn==20.1.0

Pillow==8.4.0

psycopg2==2.9.3

djongo==1.3.6
pymongo==3.12.3


requests-oauthlib==1.3.1
